+++ Rock.cpp
@@ -69,7 +69,7 @@ std::string Rock::fight(Tool* opponentTool)
 {
     // Set a temporary strength, so that the actual strength attribute is not
     // modified
-    int tempStrength = strength;
+    float tempStrength = strength;
 
     // Temporarily double this Tool's strength if the opponent is Scissors
     // Temporarily half this Tool's strength if the opponent is Paper
