+++ Scissors.cpp
@@ -69,7 +69,7 @@ std::string Scissors::fight(Tool* opponentTool)
 {
     // Set a temporary strength, so that the actual strength attribute is not
     // modified
-    int tempStrength = strength;
+    float tempStrength = strength;
 
     // Temporarily half this Tool's strength if the opponent is Rock
     // Temporarily double this Tool's strength if the opponent is Paper
